# backend/openloop/api/routes/data_sources.py
from fastapi import APIRouter, Depends, Query
from sqlalchemy.orm import Session

from backend.openloop.api.schemas import DataSourceCreate, DataSourceResponse, DataSourceUpdate
from backend.openloop.database import get_db
from backend.openloop.services import data_source_service

router = APIRouter(prefix="/api/v1/data-sources", tags=["data-sources"])


@router.post("", response_model=DataSourceResponse, status_code=201)
def create_data_source(body: DataSourceCreate, db: Session = Depends(get_db)) -> DataSourceResponse:
    ds = data_source_service.create_data_source(
        db,
        space_id=body.space_id,
        name=body.name,
        source_type=body.source_type,
        config=body.config,
        refresh_schedule=body.refresh_schedule,
    )
    return DataSourceResponse.model_validate(ds)


@router.get("", response_model=list[DataSourceResponse])
def list_data_sources(
    space_id: str | None = Query(None),
    limit: int = Query(50, le=200),
    offset: int = Query(0, ge=0),
    db: Session = Depends(get_db),
) -> list[DataSourceResponse]:
    sources = data_source_service.list_data_sources(
        db, space_id=space_id, limit=limit, offset=offset
    )
    return [DataSourceResponse.model_validate(s) for s in sources]


@router.get("/{data_source_id}", response_model=DataSourceResponse)
def get_data_source(data_source_id: str, db: Session = Depends(get_db)) -> DataSourceResponse:
    ds = data_source_service.get_data_source(db, data_source_id)
    return DataSourceResponse.model_validate(ds)


@router.patch("/{data_source_id}", response_model=DataSourceResponse)
def update_data_source(
    data_source_id: str, body: DataSourceUpdate, db: Session = Depends(get_db)
) -> DataSourceResponse:
    updates = body.model_dump(exclude_unset=True)
    ds = data_source_service.update_data_source(db, data_source_id, **updates)
    return DataSourceResponse.model_validate(ds)


@router.delete("/{data_source_id}", status_code=204)
def delete_data_source(data_source_id: str, db: Session = Depends(get_db)) -> None:
    data_source_service.delete_data_source(db, data_source_id)